% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continue_sequence.R
\name{continue_sequence}
\alias{continue_sequence}
\alias{continue_sequence,}
\alias{duplicate_last_list_element}
\title{Functions continue_sequence and duplicate_last_list_element}
\usage{
continue_sequence(chr_vector, sep = "_", factor = 1)

duplicate_last_list_element(list)
}
\arguments{
\item{chr_vector}{character, vector}

\item{sep}{character, separating prefix from suffix}

\item{factor}{numeric, increase or decrease sequence}

\item{list, }{list}
}
\value{
vector

data.frame
}
\description{
continue_sequence: suffix increase, i.e. from 1 to 2, from a to b

duplicate_last_list_element: copies last element from list and
bumps its name
}
\examples{
continue_sequence(c("a_1", "a_2"))
duplicate_last_list_element(list(a_1 = "1", b_1 = "random"))
}
\keyword{sequential}
