% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInvoice.R
\name{runInvoice}
\alias{runInvoice}
\title{Function runInvoice}
\usage{
runInvoice(installAll = FALSE)
}
\arguments{
\item{installAll}{boolean, when \code{TRUE} dependencies are installed without asking. Defaults to \code{FALSE}}
}
\value{
shiny
}
\description{
runInvoice: run shinyApp
}
\keyword{shiny}
