% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInvoice-package.R
\docType{package}
\name{shinyInvoice-package}
\alias{shinyInvoice}
\alias{shinyInvoice-package}
\title{shinyInvoice: Shiny App - Generate a Pdf Invoice with 'Rmarkdown'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generate an invoice containing a header with invoice number and businesses details. The invoice table contains any of: salary, one-liner costs, grouped costs. Under the table signature and bank account details appear. Pages are numbered when more than one. Source .json and .Rmd files are editable in the app. A .csv file with raw data can be downloaded. This package includes functions for getting exchange rates between currencies based on 'quantmod' (Ryan and Ulrich, 2023 \url{https://CRAN.R-project.org/package=quantmod}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fernandoroa/invoice-public}
  \item Report bugs at \url{https://github.com/fernandoroa/invoice-public/issues}
}

}
\author{
\strong{Maintainer}: Fernando Roa \email{froao@unal.edu.co}

}
\keyword{internal}
