% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInvoice-package.R
\docType{package}
\name{shinyInvoice-package}
\alias{shinyInvoice}
\alias{shinyInvoice-package}
\title{shinyInvoice: Shiny App. Pdf Invoice from Rmarkdown}
\description{
Generate an invoice with any of: salary, one-liner costs, grouped costs in a table. Header with invoice number, businesses details. Signature and bank account details under table. Pages numbered when more than one. Source .json and .Rmd files editable in the app. Includes functions for getting exchange rates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fernandoroa/invoice-public}
  \item Report bugs at \url{https://github.com/fernandoroa/invoice-public/issues}
}

}
\author{
\strong{Maintainer}: Fernando Roa \email{froao@ufg.br}

}
\keyword{internal}
